add rdfs:label in context and rm from node data
rdfs is sometimes causing issues when uploading files.
I guess this is because maybe sometime before an rdfs:label
was inserted into the index with a different data type.
Elastic might the have problems when the same prop with a different data type
gets added.
Since we are not using rdfs:label anywhere, we can as well remove it from the
data completly.